+++ libs/sdk-core/src/lib/db/search/search.models.ts
@@ -1,5 +1,5 @@
 import type { ExtractedDataTypes, ResourceProperties } from '../kb';
-import type { FIELD_TYPE, IResource, RelationType, RelationEntityType } from '../resource';
+import type { FIELD_TYPE, FieldId, IResource, RelationEntityType, RelationType } from '../resource';
 
 export type ResourceStatus = 'PENDING' | 'PROCESSED' | 'ERROR';
 
@@ -68,6 +68,7 @@ export namespace Search {
 
   export interface SmartResult extends IResource {
     paragraphs?: SmartParagraph[];
+    field?: FieldId;
   }
 
   export interface SmartParagraph extends Paragraph {